Organize source files
Rename `get_builtin_type_dict` to `get_type_dict`
