+++ tests/test_extend_builtins.py
@@ -5,11 +5,14 @@ import extype
 # `list` is actually the builtin type, but we import it so we get
 # proper typing. Note that the static type checker doesn't recognize
 # `[]` as an object of type list so it'll not add the extension methods.
-from extype import extend_type_with, extension, list, implement_protocol_on_type, Protocol
+
+from extype.core import implement_protocol_on_type
+from extype.extension_utils import extend_type_with, extension, Protocol
+from extype.builtin_extensions import list_ext
 
 
 def test_extend_builtins():
-    extype.extend_builtin_types()
+    list_ext.extend()
 
     assert hasattr(list, "map")
